requirements.txt
alembic==1.14.0
Flask==3.1.0
flake8==7.3.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
pymacaroons==0.13.0
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
apscheduler==3.11.0
Flask-OpenID==1.3.1
django-openid-auth==0.17 # TODO: we only need one file from this package, consider copying it to our codebase
flask-cors==5.0.0
canonicalwebteam.store-api==6.7.0
